+++ Source/Parsek/ParsekFlight.cs
@@ -7043,6 +7043,7 @@ namespace Parsek
                 longitude = lastPt.longitude,
                 altitude = lastPt.altitude,
                 rotation = lastPt.rotation,
+                rotationRecorded = true,
                 situation = inferredState == TerminalState.Splashed
                     ? SurfaceSituation.Splashed
                     : SurfaceSituation.Landed
@@ -7421,8 +7422,25 @@ namespace Parsek
                         double spawnLat, spawnLon, spawnAlt;
                         VesselSpawner.ResolveSpawnPosition(leaf, -1, lastPt,
                             out spawnLat, out spawnLon, out spawnAlt);
-                        VesselSpawner.OverrideSnapshotPosition(leaf.VesselSnapshot, spawnLat, spawnLon, spawnAlt,
-                            -1, leaf.VesselName, lastPt.rotation);
+                        if (VesselSpawner.TryResolvePreferredSpawnRotation(
+                            leaf, lastPt,
+                            out string rotationBodyName,
+                            out Quaternion rotationBodyRotation,
+                            out Quaternion surfaceRelativeRotation,
+                            out string rotationSource))
+                        {
+                            VesselSpawner.OverrideSnapshotPosition(leaf.VesselSnapshot, spawnLat, spawnLon, spawnAlt,
+                                -1, leaf.VesselName,
+                                rotationBodyName,
+                                rotationBodyRotation,
+                                surfaceRelativeRotation,
+                                rotationSource);
+                        }
+                        else
+                        {
+                            VesselSpawner.OverrideSnapshotPosition(leaf.VesselSnapshot, spawnLat, spawnLon, spawnAlt,
+                                -1, leaf.VesselName);
+                        }
                     }
 
                     uint spawnedPid = VesselSpawner.RespawnVessel(leaf.VesselSnapshot);
@@ -7653,6 +7671,7 @@ namespace Parsek
                     longitude = vessel.longitude,
                     altitude = vessel.altitude,
                     rotation = vessel.srfRelRotation,
+                    rotationRecorded = true,
                     situation = sit == Vessel.Situations.SPLASHED
                         ? SurfaceSituation.Splashed
                         : SurfaceSituation.Landed